moved event parser tests to utility tests
